+++ packages/_infra/src/analytics/edge.analytics.ts
@@ -2,7 +2,7 @@ import { Env } from '@basemaps/shared';
 import { Duration, Stack, StackProps } from 'aws-cdk-lib';
 import { Rule, Schedule } from 'aws-cdk-lib/aws-events';
 import { LambdaFunction } from 'aws-cdk-lib/aws-events-targets';
-import lf from 'aws-cdk-lib/aws-lambda';
+import lambda from 'aws-cdk-lib/aws-lambda';
 import { RetentionDays } from 'aws-cdk-lib/aws-logs';
 import { BlockPublicAccess, Bucket } from 'aws-cdk-lib/aws-s3';
 import { Construct } from 'constructs';
@@ -18,7 +18,7 @@ export interface EdgeAnalyticsProps extends StackProps {
  * Every hour create analytics based off the logs given to us from cloudwatch
  */
 export class EdgeAnalytics extends Stack {
-  public lambda: lf.Function;
+  public lambda: lambda.Function;
 
   public constructor(scope: Construct, id: string, props: EdgeAnalyticsProps) {
     super(scope, id, props);
@@ -31,12 +31,12 @@ export class EdgeAnalytics extends Stack {
       blockPublicAccess: BlockPublicAccess.BLOCK_ALL,
     });
 
-    this.lambda = new lf.Function(this, 'AnalyticLambda', {
-      runtime: lf.Runtime.NODEJS_18_X,
+    this.lambda = new lambda.Function(this, 'AnalyticLambda', {
+      runtime: lambda.Runtime.NODEJS_18_X,
       memorySize: 2048,
       timeout: Duration.minutes(10),
       handler: 'index.handler',
-      code: lf.Code.fromAsset(CODE_PATH),
+      code: lambda.Code.fromAsset(CODE_PATH),
       environment: {
         [Env.Analytics.CloudFrontId]: distributionId,
         [Env.Analytics.CacheBucket]: `s3://${cacheBucket.bucketName}`,
@@ -44,6 +44,7 @@ export class EdgeAnalytics extends Stack {
         AWS_NODEJS_CONNECTION_REUSE_ENABLED: '1',
       },
       logRetention: RetentionDays.ONE_MONTH,
+      loggingFormat: lambda.LoggingFormat.JSON,
     });
 
     cacheBucket.grantReadWrite(this.lambda);